New clientlib ver

+++ src/main/java/com/kaltura/client/services/FlavorParamsService.java
@@ -75,7 +75,7 @@ public class FlavorParamsService {
 	
 	public static class DeleteFlavorParamsBuilder extends NullRequestBuilder {
 		
-		public DeleteFlavorParamsBuilder(int id) {
+		public DeleteFlavorParamsBuilder(long id) {
 			super("flavorparams", "delete");
 			params.add("id", id);
 		}
@@ -90,13 +90,13 @@ public class FlavorParamsService {
 	 * 
 	 * @param id 
 	 */
-    public static DeleteFlavorParamsBuilder delete(int id)  {
+    public static DeleteFlavorParamsBuilder delete(long id)  {
 		return new DeleteFlavorParamsBuilder(id);
 	}
 	
 	public static class GetFlavorParamsBuilder extends RequestBuilder<FlavorParams, FlavorParams.Tokenizer, GetFlavorParamsBuilder> {
 		
-		public GetFlavorParamsBuilder(int id) {
+		public GetFlavorParamsBuilder(long id) {
 			super(FlavorParams.class, "flavorparams", "get");
 			params.add("id", id);
 		}
@@ -111,7 +111,7 @@ public class FlavorParamsService {
 	 * 
 	 * @param id 
 	 */
-    public static GetFlavorParamsBuilder get(int id)  {
+    public static GetFlavorParamsBuilder get(long id)  {
 		return new GetFlavorParamsBuilder(id);
 	}
 	
@@ -166,7 +166,7 @@ public class FlavorParamsService {
 	
 	public static class UpdateFlavorParamsBuilder extends RequestBuilder<FlavorParams, FlavorParams.Tokenizer, UpdateFlavorParamsBuilder> {
 		
-		public UpdateFlavorParamsBuilder(int id, FlavorParams flavorParams) {
+		public UpdateFlavorParamsBuilder(long id, FlavorParams flavorParams) {
 			super(FlavorParams.class, "flavorparams", "update");
 			params.add("id", id);
 			params.add("flavorParams", flavorParams);
@@ -183,7 +183,7 @@ public class FlavorParamsService {
 	 * @param id 
 	 * @param flavorParams 
 	 */
-    public static UpdateFlavorParamsBuilder update(int id, FlavorParams flavorParams)  {
+    public static UpdateFlavorParamsBuilder update(long id, FlavorParams flavorParams)  {
 		return new UpdateFlavorParamsBuilder(id, flavorParams);
 	}
 }
